Generics and Java 7 syntax in osgi package

+++ src/java/org/apache/ivy/osgi/obr/OBRResolver.java
@@ -88,7 +88,7 @@ public class OBRResolver extends AbstractOSGiResolver {
                 Resource obrResource = new URLResource(url);
                 CacheResourceOptions options = new CacheResourceOptions();
                 if (metadataTtl != null) {
-                    options.setTtl(metadataTtl.longValue());
+                    options.setTtl(metadataTtl);
                 }
                 if (forceMetadataUpdate != null) {
                     options.setForce(forceMetadataUpdate);
